Change parser and expression types from float to double for improved precision; format result display with six decimal places.

+++ include/equation.h
@@ -26,7 +26,7 @@ public:
     ~Equation() {};
 
 private:
-    exprtk::parser<float> parser;
-    exprtk::expression<float> expression;
+    exprtk::parser<double> parser;
+    exprtk::expression<double> expression;
 };
 #endif
